Make pubkey-functions more forgiving wrt CA_PASS
Particularly when generate_pubkey_pkcs11 is called, the CA_PASS
variable is not set. In many cases, it should not be set as
generate_pubkey_pkcs11 is being called from init_slot, which may
be called from a number of non-CA places.

This patch checks that the CA_PASS variable is available before
telling openssl to use it as the password. In other words, if the
environment variable is missing, openssl will ask the user for
the PIN one extra time instead of just crashing.